Upgraded Operation Enum Class
Created an Operations file that handles operations only, improve Enum Class Operations to have their character values predetermined.

Multiple functions and reoccurring statements are placed into a function so that it is easily readable.
This is so that new operators added to the operations file would be much easier to be managed.